Core change: doctor telling Xcode 4.3 w/0 CLT is experimental
Only on Lion, tell the user about the options, which are
- Xcode without CLT
- CLT without Xcode
- both (ok, it's not directly stated, but implicit)

So if no Xcode is found and we are on Lion or above,
we don't fail but check for the CLTs.

For older Macs, the old message that Xcode is needed
and the installer should be run is displayed.

If the CLT are not found but Xcode is, then we
print out about the experimental status of this setup.